Removed VS compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2675 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/items/bowling.cpp
@@ -40,7 +40,7 @@ Bowling::Bowling(Kart *kart) : Flyable(kart, POWERUP_BOWLING, 50.0f /* mass */)
     }
     else
     {
-        int min_speed = m_speed*4;
+        float min_speed = m_speed*4.0f;
         /* make it go faster when throwing forward
            so the player doesn't catch up with the ball
            and explode by touching it */
